Breaking functions into their own files
Adding cursor handling
