+++ src/hwm.cpp
@@ -54,163 +54,9 @@ hwm::hwm(QLineEdit *inFilebox, QCheckBox *inManualCheck, QComboBox *inUnitCombob
 
 }
 
-int hwm::readHWMData()
+QString hwm::getErrorString()
 {
-    QString Line;
-    QStringList List;
-    int nLines;
-
-    QFile MyFile(this->fileBox->text());
-
-    //Check if we can open the file
-    if(!MyFile.open(QIODevice::ReadOnly|QIODevice::Text))
-    {
-        this->hwmErrorString = MyFile.errorString();
-        return -1;
-    }
-
-    try
-    {
-        nLines = 0;
-        while(!MyFile.atEnd())
-        {
-            Line = MyFile.readLine();
-            nLines = nLines + 1;
-        }
-        MyFile.close();
-        MyFile.open(QIODevice::ReadOnly|QIODevice::Text);
-        this->highWaterMarks.resize(nLines);
-        int i=0;
-        while(!MyFile.atEnd())
-        {
-            Line = MyFile.readLine().simplified();
-            List = Line.split(",");
-            this->highWaterMarks[i].lon = List.value(0).toDouble();
-            this->highWaterMarks[i].lat = List.value(1).toDouble();
-            this->highWaterMarks[i].bathy = List.value(2).toDouble();
-            this->highWaterMarks[i].measured = List.value(3).toDouble();
-            this->highWaterMarks[i].modeled = List.value(4).toDouble();
-            this->highWaterMarks[i].error = List.value(5).toDouble();
-            i = i + 1;
-        }
-        return 0;
-    }
-    catch(...)
-    {
-        this->hwmErrorString = "Unexpected error reading file.";
-        return 1;
-    }
-
-}
-
-int hwm::computeLinearRegression()
-{
-    double SumXY, SumX2, SumX, SumY, SumY2, N, NDry, N2, SSE, SSTOT, YBar, SumErr, MeanErr;
-    double M,B,R2,StdDev;
-    int i;
-
-    bool ForceThroughZero = this->forceThroughZeroCheckbox->isChecked();
-
-    SumXY  = 0;
-    SumX2  = 0;
-    SumY2  = 0;
-    SumY   = 0;
-    SumX   = 0;
-    M      = 0;
-    B      = 0;
-    N      = static_cast<double>(this->highWaterMarks.size());
-    NDry   = 0;
-    SSE    = 0;
-    SSTOT  = 0;
-    SumErr = 0;
-
-    try
-    {
-        for(i = 0; i<N; i++)
-        {
-            //We ditch points that didn't wet since they
-            //skew calculation
-            if(this->highWaterMarks[i].modeled>-9999)
-            {
-                SumX   = SumX   + (this->highWaterMarks[i].measured);
-                SumY   = SumY   + (this->highWaterMarks[i].modeled);
-                SumXY  = SumXY  + (this->highWaterMarks[i].measured*this->highWaterMarks[i].modeled);
-                SumX2  = SumX2  + (this->highWaterMarks[i].measured*this->highWaterMarks[i].measured);
-                SumY2  = SumY2  + (this->highWaterMarks[i].modeled*this->highWaterMarks[i].modeled);
-                SumErr = SumErr +  this->highWaterMarks[i].error;
-            }
-            else
-                NDry = NDry + 1;
-        }
-
-        //Number of points that we'll end up using
-        N2 = N - NDry;
-
-        //Calculate the slope (M) and Correllation (R2)
-        if(ForceThroughZero)
-        {
-            //Slope
-            M   = SumXY / SumX2;
-
-            //Forced through zero
-            B   = 0;
-
-            //Average Y
-            YBar = SumY / N2;
-
-            //Calculate Total Sum of Squares
-            for(i = 0; i<N; i++)
-            {
-                //We ditch points that didn't wet since they
-                //skew calculation
-                if(this->highWaterMarks[i].modeled>-9999)
-                {
-                    SSTOT = SSTOT + qPow((this->highWaterMarks[i].modeled - YBar),2.0);
-                }
-            }
-
-            //Sum of square errors
-            SSE = SumY2 - M*M*SumX2;
-
-            //R2
-            R2  = 1 - ( SSE / SSTOT );
-        }
-        else
-        {
-            //Slope
-            M  = (N2*SumXY - SumX*SumY) / (N2*SumX2-(SumX*SumX));
-
-            //Intercept
-            B  = ((SumY*SumX2)-(SumX*SumXY))/(N2*SumX2-(SumX*SumX));
-
-            //R2 calculation
-            R2 = qPow(((N2*SumXY - (SumX*SumY)) /
-                   sqrt( (N2*SumX2 - (SumX*SumX))*(N2*SumY2-(SumY*SumY)) )),2.0);
-        }
-
-        //Calculate Standard Deviation
-        MeanErr = SumErr / N2;
-        SumErr = 0;
-        for( i = 0; i<N; i++)
-            if(this->highWaterMarks[i].modeled>-9999)
-                SumErr = SumErr + qPow(this->highWaterMarks[i].error - MeanErr,2.0);
-
-        StdDev = qSqrt(SumErr / N2);
-
-
-    }
-    catch(...)
-    {
-        return 1;
-    }
-
-    this->regLineSlope = M;
-    this->regLineIntercept = B;
-    this->regCorrelation = R2;
-    this->regStdDev = StdDev;
-
-    return 0;
-
+    return this->hwmErrorString;
 }
 
 int hwm::classifyHWM(double diff)
